Conflicts:
	plugins/tracker/include/Tracker/DateReminder/Tracker_DateReminderRenderer.class.php

+++ plugins/git/db/uninstall.sql
@@ -11,5 +11,5 @@ DELETE FROM service WHERE short_name='plugin_git';
 DELETE FROM reference_group WHERE reference_id=30;
 DELETE FROM reference WHERE id=30;
 
-DELETE FROM permissions_values WHERE permission_type IN ('PLUGIN_GIT_READ', 'PLUGIN_GIT_WRITE', 'PLUGIN_GIT_WPLUS');
+DELETE FROM permissions_values WHERE permission_type IN ('PLUGIN_GIT_READ', 'PLUGIN_GIT_WRITE', 'PLUGIN_GIT_WPLUS', 'PLUGIN_GIT_ADMIN');
